# Improve sample performance for small sets