--- a/inst/include/dqrng_sample.h
+++ b/inst/include/dqrng_sample.h
@@ -65,15 +65,6 @@ inline Rcpp::Vector<RTYPE> no_replacement_set(std::function<INT(INT)> selector, 
   return result;
 }
 
-template<typename INT>
-inline INT roulette_wheel_selection(const dqrng::rng64_t &rng, INT m, Rcpp::NumericVector probs, double max_probs) {
-  while (true) {
-    INT index = (*rng)(m);
-    if (dqrng::uniform01((*rng)()) < probs(index) / max_probs)
-      return index;
-  }
-}
-
 template<int RTYPE, typename INT>
 inline Rcpp::Vector<RTYPE> sample(const dqrng::rng64_t &rng,
                                   INT m,
@@ -85,7 +76,13 @@ inline Rcpp::Vector<RTYPE> sample(const dqrng::rng64_t &rng,
   if (probs.isNotNull()) {
     Rcpp::NumericVector tmp = probs.as();
     double max_probs = Rcpp::max(tmp);
-    selector = [rng, tmp, max_probs] (INT m) {return roulette_wheel_selection<INT>(rng, m, tmp, max_probs);};
+    selector = [rng, tmp, max_probs] (INT m) {
+      while (true) {
+        INT index = (*rng)(m);
+        if (dqrng::uniform01((*rng)()) < tmp(index) / max_probs)
+          return index;
+      }
+    };
   }
   if (replace || n <= 1) {
     return dqrng::sample::replacement<RTYPE, INT>(selector, m, n, offset);